Addition of viewAll
This viewAll function is modeled directly after the other groups as shared by Alex. I recall the discussion on whether the functions should directly output, but I think the viewAll is an exception. I thought of using printf, but I searched for similar examples and I think this is equally acceptable for this stage. I could not test with an actual field, but compiling with random arrays seem to show that it works. Please let me know if you have any thoughts or if you identify any bugs. Thank you.