viewAll Added (#34)
* Header Update #1

The header does the following:
- Select a random side to begin the game.
- Assign the number of actions left (2 in the first turn).

Please let me know if the header should do anything else, or if the script is problematic. Thank you.

* Constructor Update #2

I have applied the changes from the previous commit. The constructor does the following: 
- Randomly choose one side to begin the game
- Set victory condition to false at the beginning
- Set actions Left to 2 for the starting team

* Update TicTacToe.cpp

* Update TicTacToe.cpp

* Addition of viewAll

This viewAll function is modeled directly after the other groups as shared by Alex. I recall the discussion on whether the functions should directly output, but I think the viewAll is an exception. I thought of using printf, but I searched for similar examples and I think this is equally acceptable for this stage. I could not test with an actual field, but compiling with random arrays seem to show that it works. Please let me know if you have any thoughts or if you identify any bugs. Thank you.

* viewAll Changes 1

* viewAll Changes #2

The field is directly outputted.

* viewAll Changes #3

* Update TicTacToe.cpp

* Update TicTacToe.cpp

* Update TicTacToe.cpp

* Update TicTacToe.cpp

* Update TicTacToe.cpp